Fix form label translations.
Basic problem was that the label was being translated at declaration time - which is outside
of any request or even security/app context.
Instead, create a lazy_string which is then evaluated in the correct context.

Added tests for this.
There are other odd aspects of translating form contents itself that will need some more
work - but add a test at least so we can track it.

closes #108