Feature - implement granular Permissions.
Added support for permissions within a role (standard RBAC).

The RoleMixin provide methods for getting and modifying permissions (which
are stored as a simple comma separated string).

2 new decorators - permissions_required and permissions_accepted can be used to protect
endpoints.

A new method on UserDataStore has_permission() can be used to check if the user
has requested permission if the decorators aren't able to be used.

Permissions are simple strings and not interpreted by Flask-Security at all.

As support for this and additional functionality -introduced the idea of entire model mixins
that set up complete User and Role models for all functionality (and for future new tables).
These models also include some best practice constructs such as update_datetime and username.
These are NOT required, but are an attempt to reduce effort to integrate Flask-Security.

Added an examples/ package and implement a full example of using permissions.