fix - improve performance for auth tokens. (#168)
Turns out the flask-login will call request_loader if there is no session. That in turn
validates the token, does a DB query for the user, and verifies the contents of the token
(this is the expensive hash check).
Then - if an endpoint is decorated with either @auth_token_required or @auth_required,
it will again call request_loader and do it all over again.

This PR add a request-local flag to note that if the token has been verified and the user
looked up - and simply returns the existing user.

Also - after reading more about request and app contexts - decided that this AND the CSRF
flags (which was added a couple PRs ago) should be in the request context, not the app context.