fix - unauthorized_view doesn't accept URL. (#174)
All other config variables of type xxx_VIEW use utils.get_url() which accepts either an endpoint name or a URL.
For some reason unauthorized_view didn't. This was very confusing for folks.

Add real tests for unauthorized_view (previously they set it to "/" which is the fall-back if the code
can't find the endpoint.