2fa: fix issues with setting the tf_totp_secret during tf-setup
A new tf_totp_secret has to be created upon starting the setup process
within a session. This secret should only be committed after all checks
are successful and complete_two_factor_process is called to update the
primary_method. After that the session tf_totp_secret should be removed.