Fix post-login redirect for already authenticated users.
This was a regression when /login was changed from being decorated with @anonymous_user_required
to handling the already-authenticated case as part of the view.

That has been fixed. A subtle change in behavior is that it now uses
get_post_login_redirect() which will look for 'next' in the form or request params
and that will take precedence over POST_LOGIN_VIEW config setting. This actually makes
things more consistent, since it is EXACTLY the same now as the normal login process.

Add csrf errors to login_user.html - while this should never really occur in a real
application - it makes things much easier to debug.

Updated openapi.yaml.

closes: #221
closes: #172