Add code example for unit testing.
Added pytest unit testing to the example - to show how to mock out users and roles
and actually test an application endpoint.

During testing - realized that if permissions for a role were passed as a comma
separated string, the words weren't being stripped - that was very hard to debug.
Now - do that stripping when creating a role.

closes: #219