Integrate two-factor auth into registration and confirmation.
When two-factor auth is on and required, registration and/or confirmation will no longer
log in the user - but will redirect to 2FA setup.

Also - fixed a bug recently introduced (fixing another bug!) that didn't properly
save primary_method on first login.

Improve view_scaffold a bit to make it easy to set SECURITY configs in the environment for
easier testing.

closes: #223