Two-factor updates. (#235)
First - there are many authenticator apps out there - so change all references to google_authenticator to
simply 'authenticator' - in forms, messages, config. Tested with authy and lastpass.

Put in backwards compat code so that existing DB primary_method entries of 'google_authenticator' should still work.

Second - prior to this change - the QRcode uri was <email-username>@<service_name>. According
to the passlib docs, and consistent with other sites handling of the uri -
this should be the 'username' (e.g. full email, or username).
The service_name is sent as the issuer (as it already was). Likely this resulted when we converted to passlib.

This should have no backwards compat issues.