Introduce unified sign in.
This is a completely new implementation that leverages much of the
2FA work to enable email, sms, and authenticator app options for
signing in without a password.

Using a single form can login with identity/passcode.
Where identity is defined by USER_IDENTITY_ATTRIBUTES and the new
USER_IDENTITY_MAPPINGS used to order and match to find the appropriate
field to look up.

The passcode can be password, or a code received via email, sms, authenticator
as specified in the US_ENABLED_METHODS

Functionally complete with still a bunch of TODOs as outlined at the
top of unified_signin.py

Minor refactoring of some configuration variables that are now shared
between 2FA and unified sign in.