Add support for registering w/o a password.
Continued integrating unified sign in into other features.

We use an un-guessable password (not an empty) one in the user DB record - this makes sure
we don't let folks log in w/o any password, and keeps the DB Column definition of 'non-null' the same.

Added "password" explicitly as a US_ENABLED_METHODS - this actually allows an admin to NOT allow password login - or to JUST use passwords with this new endpoint.