Add phone number utility.
For unified signup - the user can use a phone number as an identity - so it is
important to canonicalize the input date so that minor formatting differences
don't make it impossible to look up.

Add a PhoneUtil class that uses phonenumberslite to implement validation and
canonicalization. Allow this class to be replaced as part of initialization.

Converted two-factor to use this as well for validation of input phone numbers.

Minor other documentation fixes.

closes: #250