Provide an interface so applications can catch and mitigate send_code errors.
Two-factor auth attempts to send a verification code. If that fails there was
no way for applications to intercept and perform any mitigation. Furthermore,
an exception would be thrown and a 500 returned w/o any indication of what
went wrong.

This PR fixes that by providing an overridable method in the UserMixin that
can capture and mitigate errors. The default implementation returns an error message.
The various view functions now check for errors and act accordingly.

closes: #266