Integrate freshness checking into unified sign in setup endpoint.
It became clear that just redirecting to /login was a bad user experience and
in fact doesn't actually work.

So - create 2 new endpoints: /verify and /us-verify, and backing forms that
are just for re-authenticating. default_reauthn_handler will call one or the other
based on whether UNIFIED_SIGNIN is enabled.

Added a POST_VERIFY_VIEW that can redirect upon successful re-auth. For form-based
apps - this normally isn't needed since we properly propagate the ?next=xxx parameters.

This meant the REAUTHENTICATION_VIEW was no longer needed and was removed.

Make sure json responses provide enough info when asking for re-authentication.

Fixed bug when 'within' was 0.