The alternative id feature introduced an issue with postgres/psycopg2.
Ultimately decided that we should just say - you either use the alternative id or not
and not try to fallback if one fails.

What this means is that if an app changes from user.id to alternate id (fs_uniquifier) then all
clients will be required to log in again (their session cookie won't be valid).

closes: #298