Small improvements to the freshness mechanism.
In the original code - the check for "basic" was done in the decorator - that has been moved to the utility method so callers that can't use the decorator can get identical behavior.

Added a request global fs_auth_via and set that in the decorators so that views, if need be, can get the method that was used to authenticate.

removed unused conftest endpoint and template.