Revamp how USER_IDENTITY_ATTRIBUTES is configured and used.
Aka - get rid of get_user.

Now, when looking up a user by identity attribute (such as in login, cli), FS will loop through each configured USER_IDENTITY_ATTRIBUTE and calling a mapping function which determines whether the input value is valid for that attribute - then the usermodel.find_user is called with the matching attribute name and value. This paradigm was introduced in 3.4 as part of Unified Sign In - and now is in use everywhere.

removing usermodel.get_user resolves a lot of historic issues with DBs and ORMs really not liking querying on different type entities.

Fix/enhance CLI so that you can create a user with multiple identity attributes (before the same value was set to all identity attributes).

N.B. with this PR - there is a regression with registration as noted in CHANGES.rst

closes: #337