Fix peewee
peewee ORM has been broken for a while - basically there weren't tests that tested higher level functionality. Added support for lower-case in find_user.

This uncovered a bug in permissions - add/remove in the RoleMixin are fine - but all mutating operations need to be accessible from the datastore
layer so that a db.put() can be done. So new methods - add_permission_to_role and remove_permission_from_role were added to UserDatastore.

Unrelated (but what found all these issues) - change _prepare_role_modify_args - which is just used by CLI to not take user - assume user is always an object so that that method doesn't assume 'email' is the primary user identity method.

Update quickstart for peewee to hopefully be correct.

closes: #347