tests: Skip tests requiring specific data backends if not available
Skip tests that require a specific database backend if the respective
package is not available.  This is mostly intended to make it possible
to test flask-security on Python 3.9 since Pony does not support this
Python version but there is no harm in extending it to other backends.

Technically, Flask-SQLAlchemy is still required due to other tests.