Fix security vuln - QRcodes were available via GET on /qrcode - not protected with CSRF tokens (#423)
Looking into how others do this (e.g. paypal) - the QRcode is embedded in the successful response to form POST.
So - /tf-qrcode and /us-qrcode endpoints have been completely removed.

A successful response to POST /tf-setup or /us-setup will contain the embedded QRcode in addition to the constituent parts: key, username, and issuer. This allows for manual authenticator app setup. In addition the JSON response to these POST calls also have that information.

Simpler, more secure - what's not to like.

Factored out the actual creation of the QRcode to an overridle method in the TOTP class - so application can easily change size, color ,etc.
closes: #418