Make email templates more flexible by including token in addition to link.
Email templates for registration, confirmation, reset, us_instructions now get the required one-time token in addition to the link (which has always contained the token). This allows applications - such as mobile apps - to customize the URL link rather than going directly to the Flask app.

Most of the changes are tests - turns out we really didn't check email templates.

A few signals have extra arguments passed in order to try to standardize the names of attributes sent to signals and templates.

closes: #449