Typing part 1
Get datastore and utilities typed.

Most of the issues are around the fact that the application has to provide the User/Role Model classes, and then glue that together with whichever DB implementation they want.

Since we don't know which DB implementation we provide a canned User/Role class that is just used for typing and declares the attributes as python types.

This also required a bunch of work getting sphinx working and cross-referencing other packages.

Lots more to go - still many errors etc...