Two factor improvements.
Add 'usage' to model - each webauthn security key must be designated as either a 'first' or 'secondary' key.
This makes it possible to discern which if any webauthn keys might be used as a second factor.

Improve the DB models so that from a API perspective we use lists for say extensions, and convert to comma separated as part of the model (if required).

Add support for using webauthn as 'first' authentication and using other TF for second.

Start more documentation for WebAuthn.

Refactored UserVerification out of signin and made that configurable from webauthn_util.