Introduce a two-factor plugin framework.
Now that we have 2 different two factor implementations (code and webauthn) there was a lot of duplicate code and lots of
changes to 2FA independent code to shoehorn in the new one (webauthn). With a request to add Duo 2FA it became clear that we really needed
to abstract out the common code.

Also - this provided the obvious place to put in a tf-select endpoint - used when a user has setup more than one 2FA mechanism - they have to choose which to use!

A couple other small changes:
- add simple_render_json - for endpoints that aren't returning any form data - just creates a CSRF token and returns.
- add get_form_field_xlate - make a lazy string for arbitrary text - need to use this in unified signin etc - which really aren't correctly localizable.