Webauthn fixes for Safari and Firefox
Both Safari and FF require an input element for 'focus' - add a hidden one.
Fix error message showing in the template when credential_create fails.

We weren't handling UserVerification quite correctly - now, on registration or signin start - if the app wants UserVerification, we store that in the state and properly pass that when validating the credential.

wan-delete needs to handle GET - since that happens when an endpoint needs fresh/verify and redirects back to the originating endpoint.

Not all browsers handle getTransports() - make that optional, and fixed issue with our DB model List to comma separated string when the list was empty.