Improve username support. (#625)
The LoginForm now has a username field added if USERNAME_ENABLE is set - rather than changing the type of the 'email' field
to string. This effectively once and for all removed the old flask-security hack of allowing username or email (or anything else) to be set in the email field and used as a possible identity.
The Login template now properly display input fields based on whether they correspond to a USER_IDENTITY_ATTRIBUTES.

Changed API for login to return the current list of USER_IDENTITY_ATTRIBUTES so templates/json can make good decisions.

Add an example in the docs for how to create your own LoginForm that has the old legacy behavior.

Changed the name of the higher level find_user method - it was very confusing since there is a DataStore find_user method as well and they were both used in similar places.

Finally remove encrypt_password() method - it has been deprecated for years.