Improve empty password support. (#627)
Previously, if a user registered w/o a passwd (which was allowed if UNIFIED_SIGNIN was enabled) we put in a non-guessable password in the password User record. This meant that after registration, there was no way to tell if a user had a legitimate password or not. Not great.

Now - change User DB model to allow the password field to be nullable.
Change registration and change endpoints to properly handle an empty password.

Add a config variable - PASSWORD_REQUIRED (default True) that selects whether an empty password is allowed.

Enhance change template to properly show or hide the 'current' password. Note that for users that don't have passwords, the /change endpoint is protected via a freshness check.

Fix small bug that GET /change with json didn't work.