Support for latest Werkzeug and Flask and Flask-Login
flask-login changed where they store the current user - we used that (it was internal). It is now on 'g' which is part of the app context so for
a couple unit tests - wasn't getting popped (it used to be on the request context).

Following Flask guidelines - changed our get/set of request attributes (fs_authn_via) over to 'g' as well.

Other fixes for the upcoming Flask 2.2 release.

Pin requirements to Flask<2.3 since we know we have some deprecations to worry about (@before_first_request, and JSONEncoder)