Remove deprecated add/remove_permissions methods.
The Role Object stores permissions as a list always - the underlying DB ORM implementations now
convert that to an appropriate DB type. For SQLAlchemy derived DBs - that is still a comma separated string.
There should be no DB migration required since the Column type is still UnicodeText.

Changed the AsAlist DB type to be UnicodeText so we have a unlimited size for these comma separated strings.