Updates to work with Flask >=2.2 JSON serialization changes.
Flask in 2.2 deprecated JSONEncoder and created a JSONProvider interface. In addition, Flask deprecated placing a json serializer on the blueprint (which
Flask-Security did).

Now, Flask-Security subclasses JSONProvder interface and adds its support for serializing lazy_strings. It still supports older Flask JSONEncoder interface.
Since the JSONEncoder is going away, remove the support for applications to pass in their own JSONEncoder as part of FS initialization - not sure anyone actually
did that - they can always update to Flask 2.2 and implement the JSONProvider interface.

Update pytest.ini to error out on warnings - needs a bunch of overrides right now - but is a good practice.