Remove dependence on pkg_resources (replace with importlib_resources).
pkg_resources now throws a warning that we really really shouldn't be using it.

To replace pkg_resources version parsing:
  - remove old Werkzeug version compatibility check
  - change flask json compatibility to use checks on attributes rather than version numbers. This should
    also fix the #777 (Quart compatibility) and allowed us to remove the upper bound Flask version check (which is considered bad form).

To replace pkg_resources 'files' API which we use to find Babel translations - change SECURITY_I18N_DIRNAME to default to 'builtin' which is checked and initialized as part of configuring our Flask-Babel Domain instance.

This also allows to remove the (recently added) dependence on setuptools.