A better fix for open redirect vulnerability. (#906)
In the few cases where user input is used for redirecting (via the 'next' parameter) pull apart the URL and quote the path. This will render any of
the bizarre URLs that are relative according to the spec, but interpreted as absolute by many browsers definitely relative (and harmless).

Remove the previous fix including REDIRECT_VALIDATE_MODE and REDIRECT_VALIDATE_RE configuration variables.
The Werkzeug default of sending relative URLs in Location header was restored as well.