Minor change to how LoginForm is created/modified
The default LoginForm now reflects the default configuration - email is required.
In init_app() build_login_form() is called that if USERNAME_ENABLE is set will add the username field (as before)
and change the email field to be Optional(). This is a small semantic change - prior the email field was not marked as required.

Change the new RegisterFormV2 construction - now the default form reflects the default configuration - new_password and confirm_password are required. From init_app build_register_form() is called and it will:
1) remove password_confirm field if PASSWORD_CONFIRM_REQUIRED is False
2) add username field if USERNAME_ENABLE is True
3) mark the password field as optional if PASSWORD_REQUIRED is False or UNIFIED_SIGNING is True

Simplify the PASSWORD_REQUIRED logic - before we always checked for PASSWORD_REQUIRED or UNIFIED_SIGNIN - now - at init_app() time
we throw a ValueError if PASSWORD_REQUIRED==False and UNIFIED_SIGNIN feature not enabled (which is how it has always been documented).