Add confirmation_link and reset_link to welcome_existing email template.
Improve usability by providing the email template when someone tries to register with the same email.
Both the link and token for both confirmation and resetting password are provided.
The confirmation link of course will only be provided is `confirmable` is set AND the user hasn't yet confirmed.

Note that previously - the recovery link was sent (it still is) but the reset_link is much easier and friendlier

close #1093