Support forgot password (/reset) for authenticated users also. (#1137)
For long-lived sessions it is quite possible that a user is still authenticated but has forgotten their password
and the system is requesting it for some sensitive operation (e.g /verify).
Rather than force the user to log out then start the reset process - let them start it while authenticated.

The reset flow is the same - with the end result of the user changing their password and being required to log back
in.

closes #1123