Fix localization of the tf_select SelectField. (#1140)
tf_plugin::get_setup_methods used to return a list of strings of setup methods (e.g. ["webauthn"].
This was set into the TfSelectForm as choices - which since it was just a list of string meant that WTForms used the string for BOTH the value and label.
This effectively meant we didn't really have a stable API since the label can change based on language!

Now get_setup_methods returns a tuple of (value, label).