Move to libpass and support bcrypt 5.0
We now use libpass for all python versions.

In bcrypt 5.0 they started throwing a ValueError for passwords/secrets longer than 72 bytes.
It is important to know that by default Flask-Security performs a double hash - taking the secret, using HMAC(SHA512) then
b64encodng the result.
This means that ANY password will be longer than 72 bytes (86 to be exact). In the past bcrypt would silently truncate the
input - now we have to do that explicitly.

OWASP says truncation concerns are negligible:
 https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html#input-limits-of-bcrypt

 Also bumped some minimum version for dependencies - Flask, cryptography, argon_cffi, webauthn - not strictly necessary but like to keep dependencies with in a year or so.